requirements.txt
python-dotenv==1.0.0
redis==5.0.1
pandas==2.1.1
numpy==1.26.0
mysql-connector-python==8.2.0
pywin32==306